Test CDI Bean with Arquillian
