+++ wildfly-getting-started-archetype/src/main/resources/archetype-resources/src/test/java/__defaultClassPrefix__ServiceIT.java
@@ -1,0 +1,41 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package ${package};
+
+import static org.junit.Assert.assertEquals;
+
+import jakarta.inject.Inject;
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.shrinkwrap.api.Archive;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.spec.JavaArchive;
+import org.jboss.shrinkwrap.api.spec.WebArchive;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+/**
+ * Run integration tests with Arquillian to be able to test CDI beans
+ */
+@RunWith(Arquillian.class)
+public class  ${defaultClassPrefix}ServiceIT {
+
+    @Deployment
+    public static WebArchive createTestArchive() {
+        return ShrinkWrap.create(WebArchive.class, "${defaultClassPrefix}ServiceIT.war")
+                .addClass(${defaultClassPrefix}Service.class);
+    }
+
+    @Inject
+    ${defaultClassPrefix}Service service;
+
+    @Test
+    public void testService() {
+        String result = service.hello("World");
+        assertEquals("Hello 'World'.", result);
+
+        result = service.hello("Monde");
+        assertEquals("Hello 'Monde'.", result);
+    }
+}
