--- wildfly-getting-started-archetype/src/main/resources/archetype-resources/src/test/java/__defaultClassPrefix__ServiceTest.java
@@ -1,26 +1,0 @@
-#set( $symbol_pound = '#' )
-#set( $symbol_dollar = '$' )
-#set( $symbol_escape = '\' )
-package ${package};
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
-
-/**
- * Run unit tests
- */
-public class ${defaultClassPrefix}ServiceTest {
-
-     ${defaultClassPrefix}Service service = new  ${defaultClassPrefix}Service();
-
-    @Test
-    public void testService() {
-        String result = service.hello("World");
-        assertEquals("Hello 'World'.", result);
-
-        result = service.hello("Monde");
-        assertEquals("Hello 'Monde'.", result);
-    }
-}
-
